Refactor AccessRuleService package path
The AccessRuleService class has been moved to a new package. This commit modifies the package declaration and all import statements in the codebase that referred to the AccessRuleService. This reorganization improves the structure of the code, making it easier to manage and understand.
