+++ pic-sure-auth-services/src/test/java/edu/harvard/hms/dbmi/avillach/auth/service/impl/authorization/AccessRuleServiceTest.java
@@ -2,6 +2,7 @@ package edu.harvard.hms.dbmi.avillach.auth.service.impl.authorization;
 
 import edu.harvard.hms.dbmi.avillach.auth.entity.AccessRule;
 import edu.harvard.hms.dbmi.avillach.auth.repository.AccessRuleRepository;
+import edu.harvard.hms.dbmi.avillach.auth.service.impl.AccessRuleService;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
